Handle network error
In cases of poor network reception or even no reception at all, a message is shown on the screen informing
users of network problem.

Also, Instead of using api from the env file, I added the api key
inside the weather file. This is for testing purposes for users who
might be interested in testing out the app but doesn't have account
with openweathermap.