ref: refactor
Refactored how data is being fetched so errors can be properly handled.
A class was defined to check the status of returned data so it'd be easy
to know what to do with it.
For city weather data, a model was created to extract just the needed
data from the api response.